improvement: move FavoriteIcon to core
The FavoriteIcon is located in the form directory, while it does not
have a withJarb or is FieldCompatible.

Moved the FavoriteIcon to the core directory.
