+++ src/core/FavoriteIcon/FavoriteIcon.stories.tsx
